requirements.txt
fastapi == 0.115.6   
langchain-core ==  0.3.25   
python-dotenv ==  1.0.1   
pydantic == 2.10.3    
requests ==  2.32.3   
typing-extensions ==  4.12.2    
uvicorn == 0.34.0
langchain-openai == 0.3.1